feat: loading spinner on connectionModal and when visiting profiles.
I realized you cant have skeleton loading on the modal, as the connections are all fetched when the profile is visited, as per the logic of counting the connections. Its up for discussion if thats a bit to heavy loadwise in the longrun. Anyways, I setup the spinner in its stead for now